Refactor agent.ts into modular structure and fix TypeScript errors
